move business logic to service layer

+++ src/main/java/com/restapi/usersmanagement/service/UserService.java
@@ -1,5 +1,6 @@
 package com.restapi.usersmanagement.service;
 
+import com.restapi.usersmanagement.exception.EntityNotFoundException;
 import com.restapi.usersmanagement.model.User;
 import com.restapi.usersmanagement.repository.UserRepository;
 import com.restapi.usersmanagement.repository.UserRoleRepository;
@@ -28,12 +29,18 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(int id) {
+    public void deleteUser(int id, int version) {
+        var user = findUserById(id);
+        if (user == null) {
+            logger.info(String.format("User with id: %s does not exist.", id));
+        }
+
         var userInUserRole = userRoleRepository.findUserRoleByUserId(id);
-        if (userInUserRole.isEmpty()) {
+        if (userInUserRole.isEmpty() && user.getVersion() == version) {
             userRepository.deleteById(id);
+            logger.info(String.format("User with id: %s was successfully deleted.", id));
         } else {
-            logger.info(String.format("User with id: %s exists in user_role table and therefore can't be deleted.", id));
+            logger.info(String.format("User with id: %s can not be deleted because either userRole is not empty or due to version mismatch.", id));
         }
     }
 
@@ -47,5 +54,18 @@ public class UserService {
 
     public List<User> findUsersAndUserRolesInAUnit(int unitId) {
         return userRepository.findAllByUnitIdAndUserRole(unitId);
+    }
+
+    public void updateUser(int id, User patchUser) {
+        User user = findUserById(id);
+        if (user == null) {
+            throw new EntityNotFoundException("User can not be updated because it does not exist.");
+        }
+        if (user.getVersion() == patchUser.getVersion()) {
+            save(patchUser);
+            logger.info(String.format("User with id: %s was successfully updated.", id));
+        } else {
+            logger.info(String.format("User with id: %s can not be updated because of version mismatch.", id));
+        }
     }
 }
